Сервис/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
numpy==2.0.2
Pillow==11.0.0
medpy==0.5.2
tensorflow-cpu==2.18.0
keras==3.8.0